requirements.txt
swarms==7.4.4
swarm-models
swarms-memory
requests==2.31.0
python-dotenv==1.0.1 